requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
asyncpg
sqlalchemy[asyncio]
alembic
pydantic
grpcio-reflection